+++ entrypoints/worker/index.ts
@@ -1,11 +1,12 @@
 import { sentry } from "@/sentry";
 import { getLogger } from "@logtape/logtape";
 import chalk from "chalk";
-import { getDeliveryWorker } from "~/classes/workers/delivery";
-import { getFetchWorker } from "~/classes/workers/fetch";
-import { getInboxWorker } from "~/classes/workers/inbox";
-import { getMediaWorker } from "~/classes/workers/media";
-import { getPushWorker } from "~/classes/workers/push";
+import { getDeliveryWorker } from "~/classes/queues/delivery";
+import { getFetchWorker } from "~/classes/queues/fetch";
+import { getInboxWorker } from "~/classes/queues/inbox";
+import { getMediaWorker } from "~/classes/queues/media";
+import { getPushWorker } from "~/classes/queues/push";
+import { getRelationshipWorker } from "~/classes/queues/relationships";
 
 process.on("SIGINT", () => {
     process.exit();
@@ -36,4 +37,8 @@ serverLogger.info`Starting Media Worker...`;
 getMediaWorker();
 serverLogger.info`${chalk.green("✔")} Media Worker started`;
 
-serverLogger.info`${chalk.green("✔✔✔✔✔")} All workers started`;
+serverLogger.info`Starting Relationship Worker...`;
+getRelationshipWorker();
+serverLogger.info`${chalk.green("✔")} Relationship Worker started`;
+
+serverLogger.info`${chalk.green("✔✔✔✔✔✔")} All workers started`;