+++ tests/setup.ts
@@ -3,3 +3,6 @@ import { deleteOldTestUsers } from "./utils.ts";
 
 await setupDatabase();
 await deleteOldTestUsers();
+
+// Start workers
+await import("~/entrypoints/worker/index.ts");