+++ utils/bull-board.ts
@@ -8,6 +8,7 @@ import { fetchQueue } from "~/classes/queues/fetch";
 import { inboxQueue } from "~/classes/queues/inbox";
 import { mediaQueue } from "~/classes/queues/media";
 import { pushQueue } from "~/classes/queues/push";
+import { relationshipQueue } from "~/classes/queues/relationships";
 import { config } from "~/config.ts";
 import pkg from "~/package.json";
 import type { HonoEnv } from "~/types/api";
@@ -22,6 +23,7 @@ export const applyToHono = (app: Hono<HonoEnv>): void => {
             new BullMQAdapter(fetchQueue),
             new BullMQAdapter(pushQueue),
             new BullMQAdapter(mediaQueue),
+            new BullMQAdapter(relationshipQueue),
         ],
         serverAdapter,
         options: {